requirements.txt
# Core Framework
streamlit>=1.28.0
asyncio

# AI and LLM
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=0.0.6
langgraph>=0.0.26
google-generativeai>=0.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web and APIs
requests>=2.31.0
python-dotenv>=1.0.0

# Database and Storage
sqlite3
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Utilities
tenacity>=8.2.0
cachetools>=5.3.0
Pillow>=10.0.0

# Export Functionality
reportlab>=4.0.0
python-docx>=0.8.11
markdown>=3.5.0

# Development
pytest>=7.4.0
black>=23.0.0
